+++ src/Chant/MicroPeripheral.java
@@ -1,8 +1,8 @@
 package Chant;
 
 public abstract class MicroPeripheral{
-    public byte ID;
-    public byte[] unchanged(byte cmd, byte arg1, byte arg2, byte arg3, byte arg4){return new byte[]{this.ID, cmd, arg1, arg2, arg3, arg4};}
-    public MicroPeripheral(int ID){this.ID=(byte)ID;}
-    public abstract byte[] refresh(byte cmd, byte arg1, byte arg2, byte arg3, byte arg4);
+    public short ID;
+    public short[] unchanged(short cmd, short arg1, short arg2, short arg3, short arg4){return new short[]{this.ID, cmd, arg1, arg2, arg3, arg4};}
+    public MicroPeripheral(int ID){this.ID=(short)ID;}
+    public abstract short[] refresh(short cmd, short arg1, short arg2, short arg3, short arg4);
 }
